requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
idna==3.7
Py3AMF==0.8.10
requests==2.32.3
urllib3==2.2.2
websocket-client==1.8.0
